refactor: Split up ensure.go and extract helpers
